gpu: arm: midgard: r12p0_04rel0: change temperature var type to int in model_static_power()
[jy0922.shim: apply to midgard r12p0_04rel0]

+++ drivers/gpu/arm/midgard/r12p0_04rel0/backend/gpu/mali_kbase_power_model_simple.c
@@ -36,7 +36,7 @@ static struct thermal_zone_device *gpu_tz;
 
 static unsigned long model_static_power(unsigned long voltage)
 {
-	unsigned long temperature, temp;
+	unsigned int temperature, temp;
 	unsigned long temp_squared, temp_cubed, temp_scaling_factor;
 	const unsigned long voltage_cubed = (voltage * voltage * voltage) >> 10;
